re: move in-memory SFTP server to testhelper_test
